+++ tests/Command/CommandTest.php
@@ -15,10 +15,11 @@ namespace Liip\FunctionalTestBundle\Tests\Command;
 
 use Liip\FunctionalTestBundle\Test\WebTestCase;
 use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\Console\Tester\CommandTester;
 
 class CommandTest extends WebTestCase
 {
-    private $display;
+    private $commandTester;
 
     /**
      * This method tests both the default setting of `runCommand()` and the kernel reusing, as, to reuse kernel,
@@ -27,42 +28,48 @@ class CommandTest extends WebTestCase
     public function testRunCommandWithoutOptionsAndReuseKernel(): void
     {
         // Run command without options
-        $this->display = $this->runCommand('liipfunctionaltestbundle:test');
-
-        $this->assertInternalType('string', $this->display);
+        $this->commandTester = $this->runCommand('liipfunctionaltestbundle:test');
 
         // Test default values
-        $this->assertContains('Environment: test', $this->display);
-        $this->assertContains('Verbosity level: NORMAL', $this->display);
+        $this->assertContains('Environment: test', $this->commandTester->getDisplay());
+        $this->assertContains('Verbosity level: NORMAL', $this->commandTester->getDisplay());
 
         $this->assertInternalType('boolean', $this->getDecorated());
         $this->assertTrue($this->getDecorated());
 
         // Run command and reuse kernel
-        $this->display = $this->runCommand('liipfunctionaltestbundle:test', [], true);
+        $this->commandTester = $this->runCommand('liipfunctionaltestbundle:test', [], true);
 
-        $this->assertContains('Environment: test', $this->display);
-        $this->assertContains('Verbosity level: NORMAL', $this->display);
+        $this->assertInstanceOf(CommandTester::class, $this->commandTester);
+        $this->assertEquals(0, $this->commandTester->getStatusCode());
+
+        $this->assertContains('Environment: test', $this->commandTester->getDisplay());
+        $this->assertContains('Verbosity level: NORMAL', $this->commandTester->getDisplay());
     }
 
     public function testRunCommandWithoutOptionsAndNotReuseKernel(): void
     {
         // Run command without options
-        $this->display = $this->runCommand('liipfunctionaltestbundle:test');
+        $this->commandTester = $this->runCommand('liipfunctionaltestbundle:test');
+
+        $this->assertInstanceOf(CommandTester::class, $this->commandTester);
+        $this->assertEquals(0, $this->commandTester->getStatusCode());
 
         // Test default values
-        $this->assertContains('Environment: test', $this->display);
-        $this->assertContains('Verbosity level: NORMAL', $this->display);
+        $this->assertContains('Environment: test', $this->commandTester->getDisplay());
+        $this->assertContains('Verbosity level: NORMAL', $this->commandTester->getDisplay());
 
         $this->assertInternalType('boolean', $this->getDecorated());
         $this->assertTrue($this->getDecorated());
 
         // Run command and not reuse kernel
         $this->environment = 'prod';
-        $this->display = $this->runCommand('liipfunctionaltestbundle:test', [], true);
+        $this->commandTester = $this->runCommand('liipfunctionaltestbundle:test', [], true);
 
-        $this->assertContains('Environment: prod', $this->display);
-        $this->assertContains('Verbosity level: NORMAL', $this->display);
+        $this->assertInstanceOf(CommandTester::class, $this->commandTester);
+
+        $this->assertContains('Environment: prod', $this->commandTester->getDisplay());
+        $this->assertContains('Verbosity level: NORMAL', $this->commandTester->getDisplay());
     }
 
     public function testRunCommandWithoutDecoration(): void
@@ -70,11 +77,12 @@ class CommandTest extends WebTestCase
         // Set `decorated` to false
         $this->isDecorated(false);
 
-        $this->display = $this->runCommand('liipfunctionaltestbundle:test');
+        $this->commandTester = $this->runCommand('liipfunctionaltestbundle:test');
 
-        $this->assertInternalType('string', $this->display);
+        $this->assertInstanceOf(CommandTester::class, $this->commandTester);
+        $this->assertEquals(0, $this->commandTester->getStatusCode());
 
-        $this->assertContains('Verbosity level: NORMAL', $this->display);
+        $this->assertContains('Verbosity level: NORMAL', $this->commandTester->getDisplay());
 
         $this->assertInternalType('boolean', $this->getDecorated());
         $this->assertFalse($this->getDecorated());
@@ -89,15 +97,16 @@ class CommandTest extends WebTestCase
         $this->assertInternalType('boolean', $this->getDecorated());
         $this->assertFalse($this->getDecorated());
 
-        $this->display = $this->runCommand('liipfunctionaltestbundle:test');
+        $this->commandTester = $this->runCommand('liipfunctionaltestbundle:test');
 
-        $this->assertInternalType('string', $this->display);
+        $this->assertInstanceOf(CommandTester::class, $this->commandTester);
+        $this->assertEquals(0, $this->commandTester->getStatusCode());
 
-        $this->assertEmpty($this->display);
-        $this->assertNotContains('Verbosity level: NORMAL', $this->display);
-        $this->assertNotContains('Verbosity level: VERBOSE', $this->display);
-        $this->assertNotContains('Verbosity level: VERY_VERBOSE', $this->display);
-        $this->assertNotContains('Verbosity level: DEBUG', $this->display);
+        $this->assertEmpty($this->commandTester->getDisplay());
+        $this->assertNotContains('Verbosity level: NORMAL', $this->commandTester->getDisplay());
+        $this->assertNotContains('Verbosity level: VERBOSE', $this->commandTester->getDisplay());
+        $this->assertNotContains('Verbosity level: VERY_VERBOSE', $this->commandTester->getDisplay());
+        $this->assertNotContains('Verbosity level: DEBUG', $this->commandTester->getDisplay());
     }
 
     public function testRunCommandVerbosityImplicitlyNormal(): void
@@ -109,16 +118,15 @@ class CommandTest extends WebTestCase
         $this->assertInternalType('boolean', $this->getDecorated());
         $this->assertFalse($this->getDecorated());
 
-        $this->display = $this->runCommand('liipfunctionaltestbundle:test');
+        $this->commandTester = $this->runCommand('liipfunctionaltestbundle:test');
+        $this->assertEquals(0, $this->commandTester->getStatusCode());
 
-        $this->assertContains('Verbosity level: NORMAL', $this->display);
+        $this->assertInstanceOf(CommandTester::class, $this->commandTester);
 
-        $this->assertInternalType('string', $this->display);
-
-        $this->assertNotContains('Verbosity level: VERBOSE', $this->display);
-
-        $this->assertNotContains('Verbosity level: VERY_VERBOSE', $this->display);
-        $this->assertNotContains('Verbosity level: DEBUG', $this->display);
+        $this->assertContains('Verbosity level: NORMAL', $this->commandTester->getDisplay());
+        $this->assertNotContains('Verbosity level: VERBOSE', $this->commandTester->getDisplay());
+        $this->assertNotContains('Verbosity level: VERY_VERBOSE', $this->commandTester->getDisplay());
+        $this->assertNotContains('Verbosity level: DEBUG', $this->commandTester->getDisplay());
     }
 
     public function testRunCommandVerbosityExplicitlyNormal(): void
@@ -127,16 +135,15 @@ class CommandTest extends WebTestCase
         $this->assertSame(OutputInterface::VERBOSITY_NORMAL, $this->getVerbosityLevel());
 
         $this->isDecorated(false);
-        $this->display = $this->runCommand('liipfunctionaltestbundle:test');
+        $this->commandTester = $this->runCommand('liipfunctionaltestbundle:test');
+        $this->assertEquals(0, $this->commandTester->getStatusCode());
 
-        $this->assertContains('Verbosity level: NORMAL', $this->display);
+        $this->assertInstanceOf(CommandTester::class, $this->commandTester);
 
-        $this->assertInternalType('string', $this->display);
-
-        $this->assertNotContains('Verbosity level: VERBOSE', $this->display);
-
-        $this->assertNotContains('Verbosity level: VERY_VERBOSE', $this->display);
-        $this->assertNotContains('Verbosity level: DEBUG', $this->display);
+        $this->assertContains('Verbosity level: NORMAL', $this->commandTester->getDisplay());
+        $this->assertNotContains('Verbosity level: VERBOSE', $this->commandTester->getDisplay());
+        $this->assertNotContains('Verbosity level: VERY_VERBOSE', $this->commandTester->getDisplay());
+        $this->assertNotContains('Verbosity level: DEBUG', $this->commandTester->getDisplay());
     }
 
     public function testRunCommandVerbosityVerbose(): void
@@ -144,14 +151,15 @@ class CommandTest extends WebTestCase
         $this->setVerbosityLevel('verbose');
         $this->assertSame(OutputInterface::VERBOSITY_VERBOSE, $this->getVerbosityLevel());
 
-        $this->display = $this->runCommand('liipfunctionaltestbundle:test');
+        $this->commandTester = $this->runCommand('liipfunctionaltestbundle:test');
+        $this->assertEquals(0, $this->commandTester->getStatusCode());
 
-        $this->assertInternalType('string', $this->display);
+        $this->assertInstanceOf(CommandTester::class, $this->commandTester);
 
-        $this->assertContains('Verbosity level: NORMAL', $this->display);
-        $this->assertContains('Verbosity level: VERBOSE', $this->display);
-        $this->assertNotContains('Verbosity level: VERY_VERBOSE', $this->display);
-        $this->assertNotContains('Verbosity level: DEBUG', $this->display);
+        $this->assertContains('Verbosity level: NORMAL', $this->commandTester->getDisplay());
+        $this->assertContains('Verbosity level: VERBOSE', $this->commandTester->getDisplay());
+        $this->assertNotContains('Verbosity level: VERY_VERBOSE', $this->commandTester->getDisplay());
+        $this->assertNotContains('Verbosity level: DEBUG', $this->commandTester->getDisplay());
     }
 
     public function testRunCommandVerbosityVeryVerbose(): void
@@ -163,14 +171,15 @@ class CommandTest extends WebTestCase
         $this->assertInternalType('boolean', $this->getDecorated());
         $this->assertFalse($this->getDecorated());
 
-        $this->display = $this->runCommand('liipfunctionaltestbundle:test');
+        $this->commandTester = $this->runCommand('liipfunctionaltestbundle:test');
+        $this->assertEquals(0, $this->commandTester->getStatusCode());
 
-        $this->assertInternalType('string', $this->display);
+        $this->assertInstanceOf(CommandTester::class, $this->commandTester);
 
-        $this->assertContains('Verbosity level: NORMAL', $this->display);
-        $this->assertContains('Verbosity level: VERBOSE', $this->display);
-        $this->assertContains('Verbosity level: VERY_VERBOSE', $this->display);
-        $this->assertNotContains('Verbosity level: DEBUG', $this->display);
+        $this->assertContains('Verbosity level: NORMAL', $this->commandTester->getDisplay());
+        $this->assertContains('Verbosity level: VERBOSE', $this->commandTester->getDisplay());
+        $this->assertContains('Verbosity level: VERY_VERBOSE', $this->commandTester->getDisplay());
+        $this->assertNotContains('Verbosity level: DEBUG', $this->commandTester->getDisplay());
     }
 
     public function testRunCommandVerbosityDebug(): void
@@ -182,14 +191,24 @@ class CommandTest extends WebTestCase
         $this->assertInternalType('boolean', $this->getDecorated());
         $this->assertFalse($this->getDecorated());
 
-        $this->display = $this->runCommand('liipfunctionaltestbundle:test');
+        $this->commandTester = $this->runCommand('liipfunctionaltestbundle:test');
+        $this->assertEquals(0, $this->commandTester->getStatusCode());
 
-        $this->assertInternalType('string', $this->display);
+        $this->assertInstanceOf(CommandTester::class, $this->commandTester);
 
-        $this->assertContains('Verbosity level: NORMAL', $this->display);
-        $this->assertContains('Verbosity level: VERBOSE', $this->display);
-        $this->assertContains('Verbosity level: VERY_VERBOSE', $this->display);
-        $this->assertContains('Verbosity level: DEBUG', $this->display);
+        $this->assertContains('Verbosity level: NORMAL', $this->commandTester->getDisplay());
+        $this->assertContains('Verbosity level: VERBOSE', $this->commandTester->getDisplay());
+        $this->assertContains('Verbosity level: VERY_VERBOSE', $this->commandTester->getDisplay());
+        $this->assertContains('Verbosity level: DEBUG', $this->commandTester->getDisplay());
+    }
+
+    public function testRunCommandStatusCode(): void
+    {
+        $this->commandTester = $this->runCommand('liipfunctionaltestbundle:test-status-code');
+
+        $this->assertInstanceOf(CommandTester::class, $this->commandTester);
+
+        $this->assertEquals(10, $this->commandTester->getStatusCode());
     }
 
     /**
@@ -199,13 +218,13 @@ class CommandTest extends WebTestCase
     {
         $this->setVerbosityLevel('foobar');
 
-        $this->runCommand('command:test');
+        $this->runCommand('liipfunctionaltestbundle:test');
     }
 
     public function tearDown(): void
     {
         parent::tearDown();
 
-        unset($this->display);
+        unset($this->commandTester);
     }
 }